Created Project
I created the project in Unity 2019.7. I also remove the Test.txt file as it was intended to be a temporary file to test the functionality of hte repository.
This is a completely default project with the 2D template, however it has some additional empty folders that should be apropriately populated with matching files.